POC (WIP): Inspect docker container
Note: This is not meant to be merged or even to be used as a base for
further development. It's just something I hacked together while trying
to find out how well running docker containers could be inspected.

I adapted the localhost inspection to actually inspect the 'opensuse'
container which is expected to be running. When we actually implement
that it should be moved to a separate class/command.

Working:
  * Inspection of all scopes
  * Extraction of flat file scopes

Not Working:
  * Extraction of unmanaged files (the according tarballs are created
    but seem to be empty)

TBD:
  * Figure out if docker VOLUMEs are handled well (e.g. as remote
    mounts)